headerSlice => menuSlice refactor
